+++ doc/LUA_SUPPORT.md
@@ -257,10 +257,11 @@ If you want to add new bindings, consider looking at existing examples in
   the functions there are no-op ).
 * `catalua_sol.h` and `catalua_sol_fwd.h` - Wrappers for `sol/sol.hpp` and
   `sol/forward.hpp` with custom pragmas to make them compile.
-* `catalua_bindings.h`(`.cpp`) - Game Lua bindings live here. The header file contains `luna` wrapper.
+* `catalua_bindings.h`(`.cpp`) - Game Lua bindings live here.
 * `catalua_console.h`(`.cpp`) - Ingame Lua console.
 * `catalua_impl.h`(`.cpp`) - Implementation details for `catalua.h`(`.cpp`).
-* `catalua_iuse_actor.h`(`.cpp`) - Lua-driven `iuse_actor`
+* `catalua_iuse_actor.h`(`.cpp`) - Lua-driven `iuse_actor`.
 * `catalua_log.h`(`.cpp`) - In-memory logging for the console.
-* `catalua_readonly.h`(`.cpp`) - Functions for marking Lua tables as read-only
+* `catalua_luna.h` - Usertype registration interface with automatic doc generation, aka `luna`.
+* `catalua_readonly.h`(`.cpp`) - Functions for marking Lua tables as read-only.
 * `catalua_serde.h`(`.cpp`) - Lua table to/from JSON (de-)serialization.
