+++ src/catalua_impl.h
@@ -25,6 +25,4 @@ void run_lua_script( sol::state &lua, const std::string &script_name );
 void run_console_input( sol::state &lua, const std::string &chunk );
 void check_func_result( sol::protected_function_result &res );
 
-void reg_all_bindings( sol::state &lua );
-
 #endif // CATA_SRC_CATALUA_IMPL_H
